add handlers for document type items
