+++ src/mocks/browser-handlers.ts
@@ -1,6 +1,6 @@
 import { handlers as dataTypeHandlers } from './handlers/data-type/index.js';
 import { handlers as relationTypeHandlers } from './handlers/relation-type.handlers.js';
-import { handlers as documentTypeHandlers } from './handlers/document-type.handlers.js';
+import { handlers as documentTypeHandlers } from './handlers/document-type/index.js';
 import { handlers as installHandlers } from './handlers/install.handlers.js';
 import * as manifestsHandlers from './handlers/manifests.handlers.js';
 import { handlers as publishedStatusHandlers } from './handlers/published-status.handlers.js';
