+++ src/mocks/handlers/document-type/slug.ts
@@ -1,0 +1,1 @@
+export const slug = '/document-type';
