+++ src/mocks/handlers/document-type/tree.handlers.ts
@@ -1,0 +1,29 @@
+const { rest } = window.MockServiceWorker;
+import { umbDocumentTypeData } from '../../data/document-type.data.js';
+import { slug } from './slug.js';
+import { umbracoPath } from '@umbraco-cms/backoffice/utils';
+
+export const handlers = [
+	rest.get(umbracoPath(`/tree${slug}/root`), (req, res, ctx) => {
+		const rootItems = umbDocumentTypeData.getTreeRoot();
+		const response = {
+			total: rootItems.length,
+			items: rootItems,
+		};
+		return res(ctx.status(200), ctx.json(response));
+	}),
+
+	rest.get(umbracoPath(`/tree${slug}/children`), (req, res, ctx) => {
+		const parentId = req.url.searchParams.get('parentId');
+		if (!parentId) return;
+
+		const children = umbDocumentTypeData.getTreeItemChildren(parentId);
+
+		const response = {
+			total: children.length,
+			items: children,
+		};
+
+		return res(ctx.status(200), ctx.json(response));
+	}),
+];
